+++ Userland/Libraries/LibJS/Runtime/Intl/PluralRulesConstructor.h
@@ -27,4 +27,6 @@ private:
     JS_DECLARE_NATIVE_FUNCTION(supported_locales_of);
 };
 
+ThrowCompletionOr<PluralRules*> initialize_plural_rules(GlobalObject& global_object, PluralRules& plural_rules, Value locales_value, Value options_value);
+
 }
